+++ app/_custom/config.json
@@ -1,0 +1,4 @@
+{
+   "domain": "",
+   "siteId": ""
+}
