--- app/_custom/config.ts
@@ -1,5 +1,0 @@
-type config = {
-   siteId?: string;
-   domain?: string;
-};
-export const config: config = {};
